+++ lib/misty/openstack/cinder/v1.rb
@@ -1,21 +1,22 @@
-require 'misty/http/client'
 require 'misty/openstack/cinder/cinder_v1'
+require 'misty/client_pack'
 
 module Misty
   module Openstack
     module Cinder
-      class V1 < Misty::HTTP::Client
+      class V1
         extend Misty::Openstack::CinderV1
+        include Misty::ClientPack
 
-        def self.api
-          v1
+        def api
+          self.class.v1
         end
 
-        def self.prefix_path_to_ignore
+        def prefix_path_to_ignore
           '/v1/{admin_tenant_id}/'
         end
 
-        def self.service_names
+        def service_names
           %w{block-storage, block-store, volume}
         end
       end
