+++ lib/misty/openstack/heat/v1.rb
@@ -1,21 +1,22 @@
-require 'misty/http/client'
 require 'misty/openstack/heat/heat_v1'
+require 'misty/client_pack'
 
 module Misty
   module Openstack
     module Heat
-      class V1 < Misty::HTTP::Client
+      class V1
         extend Misty::Openstack::HeatV1
+        include Misty::ClientPack
 
-        def self.api
-          v1
+        def api
+          self.class.v1
         end
 
-        def self.prefix_path_to_ignore
+        def prefix_path_to_ignore
           '/v1/{tenant_id}'
         end
 
-        def self.service_names
+        def service_names
           %w{orchestration}
         end
       end
