+++ lib/misty/openstack/ironic/v1.rb
@@ -1,19 +1,20 @@
-require 'misty/http/client'
-require 'misty/microversion'
 require 'misty/openstack/ironic/ironic_v1'
+require 'misty/client_pack'
+require 'misty/microversion'
 
 module Misty
   module Openstack
     module Ironic
-      class V1 < Misty::HTTP::Client
+      class V1
         extend Misty::Openstack::IronicV1
+        include Misty::ClientPack
         include Misty::Microversion
 
-        def self.api
-          v1
+        def api
+          self.class.v1
         end
 
-        def self.service_names
+        def service_names
           %w{baremetal}
         end
       end
