+++ lib/misty/openstack/keystone/v2_0.rb
@@ -1,20 +1,21 @@
-require 'misty/http/client'
 require 'misty/openstack/keystone/keystone_v2_0'
 require 'misty/openstack/keystone/keystone_v2_0_ext'
+require 'misty/client_pack'
 
 module Misty
   module Openstack
     module Keystone
-      class V2_0 < Misty::HTTP::Client
+      class V2_0
         extend Misty::Openstack::KeystoneV2_0
+        include Misty::ClientPack
 
-        def self.api
-          api = v2_0
-          api.merge!(v2_0_ext)
-          api
+        def api
+          all = self.class.v2_0
+          all.merge!(self.class.v2_0_ext)
+          all
         end
 
-        def self.service_names
+        def service_names
           %w{identity}
         end
       end
