+++ lib/misty/openstack/keystone/v3.rb
@@ -1,20 +1,21 @@
-require 'misty/http/client'
 require 'misty/openstack/keystone/keystone_v3'
 require 'misty/openstack/keystone/keystone_v3_ext'
+require 'misty/client_pack'
 
 module Misty
   module Openstack
     module Keystone
-      class V3 < Misty::HTTP::Client
+      class V3
         extend Misty::Openstack::KeystoneV3
+        include Misty::ClientPack
 
-        def self.api
-          api = v3
-          api.merge!(v3_ext)
-          api
+        def api
+          all = self.class.v3
+          all.merge!(self.class.v3_ext)
+          all
         end
 
-        def self.service_names
+        def service_names
           %w{identity}
         end
       end
