+++ lib/misty/openstack/magnum/v1.rb
@@ -1,19 +1,20 @@
-require 'misty/http/client'
-require 'misty/openstack/microversion'
 require 'misty/openstack/magnum/magnum_v1'
+require 'misty/client_pack'
+require 'misty/openstack/microversion'
 
 module Misty
   module Openstack
     module Magnum
-      class V1 < Misty::HTTP::Client
+      class V1
         extend Misty::Openstack::MagnumV1
+        include Misty::ClientPack
         include Misty::HTTP::Microversion
 
-        def self.api
-          v1
+        def api
+          self.class.v1
         end
 
-        def self.service_names
+        def service_names
           %w{container}
         end
       end
