+++ lib/misty/openstack/neutron/v2_0.rb
@@ -1,17 +1,18 @@
-require 'misty/http/client'
 require 'misty/openstack/neutron/neutron_v2_0'
+require 'misty/client_pack'
 
 module Misty
   module Openstack
     module Neutron
-      class V2_0 < Misty::HTTP::Client
+      class V2_0
         extend Misty::Openstack::NeutronV2_0
+        include Misty::ClientPack
 
-        def self.api
-          v2_0
+        def api
+          self.class.v2_0
         end
 
-        def self.service_names
+        def service_names
           %w{network networking}
         end
       end
