+++ lib/misty/openstack/zaqar/v2.rb
@@ -1,17 +1,18 @@
-require 'misty/http/client'
 require 'misty/openstack/zaqar/zaqar_v2'
+require 'misty/client_pack'
 
 module Misty
   module Openstack
     module Zaqar
-      class V2 < Misty::HTTP::Client
+      class V2
         extend Misty::Openstack::ZaqarV2
+        include Misty::ClientPack
 
-        def self.api
-          v2
+        def api
+          self.class.v2
         end
 
-        def self.service_names
+        def service_names
           %w{messaging}
         end
       end
